requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==24.3.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.9.0
blockdiag==3.0.0
build==1.3.0
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
cftime==1.6.3
charset-normalizer==3.3.2
chromadb==1.0.15
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
durationpy==0.10
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.8
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.47.0
frozenlist==1.5.0
fsspec==2025.3.2
funcparserlib==2.0.0a0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.74.0
h11==0.16.0
h5py==3.10.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.6
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.5
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.11
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
narwhals==2.0.1
netCDF4==1.6.5
networkx==3.4.2
numpy==1.26.2
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.98.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
overrides==7.7.0
packaging==23.2
pandas==2.1.4
passlib==1.7.4
pika==1.3.2
Pillow==10.1.0
posthog==5.4.0
propcache==0.2.1
protobuf==6.31.1
psycopg2==2.9.10
pvlib==0.7.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.1.1
pypdf==5.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.8.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
sentence-transformers==4.1.0
shellingham==1.5.4
siphon==0.9
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.45.3
streamlit==1.47.1
structlog==24.4.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.51.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.1.0
uv==0.6.12
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.5
webcolors==1.13
websocket-client==1.8.0
websockets==15.0.1
yarl==1.18.3
zipp==3.23.0
zstandard==0.23.0
